django_captcha/requirements.txt
Django==1.6
Pillow==2.6.1
South==1.0
django-simple-captcha==0.4.2
six==1.8.0
